renamed service.h to service_description.h
