+++ src/anh/service/service_directory_unittest.cc
@@ -24,7 +24,7 @@
 
 #include "anh/service/galaxy.h"
 #include "anh/service/datastore.h"
-#include "anh/service/service.h"
+#include "anh/service/service_description.h"
 #include "anh/service/service_directory.h"
 #include <anh/event_dispatcher/event_dispatcher.h>
